added content for the shift page. added id element in the get request for the users card. added link from shift page to home page
