+++ views/index.handlebars
@@ -26,7 +26,7 @@
         <div class="collapse navbar-collapse" id="navbarNav">
           <ul class="navbar-nav">
             <li class="nav-item">
-              <a class="nav-link active" aria-current="page" href="#">Home</a>
+              <a class="nav-link active" aria-current="page" href="/">Home</a>
             </li>
             <li class="nav-item">
               <a class="nav-link" href="/projects">Projects</a>
@@ -49,7 +49,7 @@
             </div>
               <ul class="list-group list-group-flush">
                 {{#each users}}
-                <li class="list-group-item">{{firstName}} {{lastName}}</li>
+                <li data-id={{id}} class="list-group-item">{{firstName}} {{lastName}}</li>
                 {{/each}}
               </ul>
   <!-- START Modal -->
